+++ backend/LaptopGG/src/main/kotlin/going9/laptopgg/dto/request/DisplayRequest.kt
@@ -12,7 +12,7 @@ data class DisplayRequest(
     var colorAccuracy: ColorAccuracy = ColorAccuracy.GOOD,  // 색 정확도
     var refreshRate: Int = 0,  // 주사율 (Hz)
     var glareType: GlareType = GlareType.GLARE,  // 글레어 유형 (NONE, GLARE, ANTI_GLARE)
-    var screenSize: Double = 0.0,  // 화면 크기 (인치 단위, 예: 15.6, 16 등)
+    var screenSize: Int = 0,  // 화면 크기 (인치 단위, 예: 15.6, 16 등)
     var isTouch: Boolean = false,  // 터치 스크린 여부
     var aspectRatio: String = "",  // 화면 비율 (예: "16:9", "16:10")
 ) {
